Updated q5
q5 currently is passing with 12/18. There may be some ways to make it faster and currently its only failing by having a non perfect win rate. it will lose 1 or 2 gaves on avrage. could use to change the pelet eating evaluation but I am not sure of a way to make it better